refactor: separate game editor state and UI
